Update
Think I found the mistake. I accidentally commented out RecBurstFlag = False, which could have been causing it to be written every time through the loop(?). Also, started working on StdDev a little, not sure I understand how to use the StdDevSpa, but it seems like the StdDev in the DataTable is working correctly