Update
Median works again. Table needs to be called EVERY time in the loop. It stores each value in its memory until the Output trigger is True (RecBurstFlag, in this case). I set that true if max values are reached AND if we exit the loop because we missed our max. I also record how many values were collected before exiting, then NAN the values in the burst table. I think. now to test.