Comms mode updated and setperiod change
Most of the code written to fix the comms mode logic problem and to set the period for cr1000 burst. Need to further test if we should use fromt 0 or 19 all the time or switch based on comms mode. Period of 1200 seems to work well. 1100 was too fast and had missing data. Currently testing 1150.