OneMinConsistent
Added in OneMinConsistent variable. When set to true it will make the 1min table write until disabled. The recCalibData will not shit it off. The PT3 subroutine will not operate so as to not lose a data point. Tripled the card allocation size to handle the larger ammount of data. Expanded the USR file to include this variable to ensure it remains correct in the event of a power cycle.