Uplooker USR update and TopBottom WQ bug fix
Added UseDir to USR drive if we are using an uplooker and autorange. If we use an uplooker without autorange, this will need to be altered. If logger powered down, UseDir would revert to false. Now it will save user selection.

Fixed bug where with a topBottom sonde configuration, if the top sonde was removed/burst was failing, the bottom sonde would NAN all median value (yet keep min, max, stddev). We were using the top sonde's BurstMissed bad data indicator in the btmburst table. Switched to use BtmBurstMissed instead.